+++ src/molecules/TabHeader.jsx
@@ -5,19 +5,23 @@ import Spacing from '../atoms/Spacing';
 import Icon from '../atoms/Icon';
 import Text from '../atoms/Text';
 
-const TabHeader = ({ isActive, onSelect, iconType, label }) => {
+const TabHeader = ({ isActive, onSelect, iconType, label, flexGrow }) => {
   const activeStyle = css`
     border-bottom: 2px solid #cf348c;
     padding: 10px 20px;
     display: flex;
     cursor: pointer;
     font-weight: 700;
+    flex: ${flexGrow ? '1' : null};
+    justify-content: ${flexGrow ? 'center' : null};
   `;
 
   const normalStyles = css`
     padding: 10px 20px;
     display: flex;
     cursor: pointer;
+    flex: ${flexGrow ? '1' : null};
+    justify-content: ${flexGrow ? 'center' : null};
   `;
 
   return (
@@ -41,10 +45,13 @@ TabHeader.propTypes = {
   iconType: PropTypes.string,
   onSelect: PropTypes.func.isRequired,
   isActive: PropTypes.bool.isRequired,
+  /** Let all the flexible items be the same length adding `flex-grow: 1` */
+  flexGrow: PropTypes.bool,
 };
 
 TabHeader.defaultProps = {
   iconType: '',
+  flexGrow: false,
 };
 
 export default TabHeader;