+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
-## v0.9.2
+## v0.9.3
 
-- Adds color variants to Link component
+- Added fixedHeight prop to keep fixed height in modal and hasPadding to remove padding.
+- Adjusted the padding for the modal title.
+- Added flex prop for tabs
+- Added max-width prop for iconText